pyproject.toml
[project]
name = "browser-use-trace-lab"
version = "0.1.0"
description = "Standalone browser-use experiment runner with OttoAuth-compatible trace output"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "browser-use>=0.1.45",
  "langchain-openai>=0.2.14",
  "openai>=1.40.0",
  "playwright>=1.45.0",
  "python-dotenv>=1.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
